+++ packages/admin/admin-vite-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @medusajs/admin-vite-plugin
 
+## 2.15.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/admin-shared@2.15.4
+
+## 2.15.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/admin-shared@2.15.3
+
 ## 2.15.2
 
 ### Patch Changes