+++ packages/cli/oas/oas-github-ci/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @medusajs/oas-github-ci
 
+## 2.15.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/medusa-oas-cli@2.15.4
+
+## 2.15.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @medusajs/medusa-oas-cli@2.15.3
+
 ## 2.15.2
 
 ### Patch Changes